+++ DependencyInjection/MassiveSearchExtension.php
@@ -35,6 +35,7 @@ class MassiveSearchExtension extends Extension
 
         $container->setAlias('massive_search.factory', $config['services']['factory']);
 
+        $loader->load('command.xml');
         $this->loadSearch($config, $loader, $container);
         $this->loadMetadata($config['metadata'], $loader, $container);
         $this->loadPersistence($config['persistence'], $loader);
@@ -51,7 +52,7 @@ class MassiveSearchExtension extends Extension
         }
     }
 
-    private function loadSearch($config, $loader, $container)
+    private function loadSearch($config, Loader\XmlFileLoader $loader, ContainerBuilder $container)
     {
         $container->setAlias('massive_search.adapter', 'massive_search.adapter.' . $config['adapter']);
         $loader->load('search.xml');
@@ -66,7 +67,7 @@ class MassiveSearchExtension extends Extension
         }
     }
 
-    private function loadZendSearch($config, $loader, $container)
+    private function loadZendSearch($config, Loader\XmlFileLoader $loader, ContainerBuilder $container)
     {
         $container->setParameter('massive_search.adapter.zend_lucene.basepath', $config['basepath']);
         $container->setParameter('massive_search.adapter.zend_lucene.hide_index_exception', $config['hide_index_exception']);
@@ -74,7 +75,7 @@ class MassiveSearchExtension extends Extension
         $loader->load('adapter_zendlucene.xml');
     }
 
-    private function loadElasticSearch($config, $loader, $container)
+    private function loadElasticSearch($config, Loader\XmlFileLoader $loader, ContainerBuilder $container)
     {
         $container->setParameter('massive_search.adapter.elastic.hosts', $config['hosts']);
         $loader->load('adapter_elastic.xml');
@@ -86,7 +87,7 @@ class MassiveSearchExtension extends Extension
         }
     }
 
-    private function loadMetadata($config, $loader, $container)
+    private function loadMetadata($config, Loader\XmlFileLoader $loader, ContainerBuilder $container)
     {
         $dir = $container->getParameterBag()->resolveValue($config['cache_dir']);
         if (!file_exists($dir)) {